refactor(store): decomposition action types
